% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.VPdtw.R
\name{plot.VPdtw}
\alias{plot.VPdtw}
\title{Plot VPdtw}
\usage{
\method{plot}{VPdtw}(
  x,
  type = c("All", "Before", "After", "Shift", "Chromatograms"),
  xlim = NULL,
  ...
)
}
\arguments{
\item{x}{A VPdtw object generated by \code{VPdtw}.}

\item{type}{What to plot.}

\item{xlim}{Numeric vector specifying x-axis limits.}

\item{...}{Additional arguments}
}
\description{
Plot VPdtw
}
