% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VPdtw.R
\name{print.VPdtw}
\alias{print.VPdtw}
\title{Print VPdtw}
\usage{
\method{print}{VPdtw}(x, ...)
}
\arguments{
\item{x}{A VPdtw object generated by \code{VPdtw}.}

\item{...}{Additional argument.}
}
\value{
Numeric vector from the \code{summary} slot in the \code{VPdtw}
object specified by \code{x}.
}
\description{
Print VPdtw
}
