% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VPdtw.R
\name{print.VPdtw}
\alias{print.VPdtw}
\title{Print VPdtw}
\usage{
\method{print}{VPdtw}(x, ...)
}
\arguments{
\item{x}{A VPdtw object generated by \code{VPdtw}.}

\item{...}{Additional argument.}
}
\description{
Print VPdtw
}
