% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VPdtw.R
\name{print.VPdtw}
\alias{print.VPdtw}
\title{Print VPdtw}
\usage{
\method{print}{VPdtw}(x, ...)
}
\arguments{
\item{x}{A VPdtw object generated by \code{VPdtw}.}

\item{...}{Additional argument.}
}
\value{
Numeric vector from the \code{summary} slot in the \code{VPdtw}
object specified by \code{x}.
}
\description{
Print VPdtw
}
\examples{
query <- c(1,5,4,3,9,8,5,2,6,5,4)
reference <- c(rnorm(5), query, rnorm(5))
lambda <- rep(0, length(reference))
maxshift <- 11
res <- VPdtw(reference, query, lambda, maxshift)
print(res)

}
